requirements.txt
aiohttp==3.8.3
aiomysql==0.1.1
aiopg==1.4.0
aiosignal==1.3.1
aiosqlite==0.17.0
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
beautifulsoup4==4.11.1
braceexpand==0.1.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cookies-discord-components==3.0.0
discord-components-mirror==2.1.5
discord-ext-bot==1.0.1
discord-ext-forms==3.0.3
discord-ext-menus @ git+https://github.com/Rapptz/discord-ext-menus@8686b5d1bbc1d3c862292eb436ab630d6e9c9b53
discord-py-interactions==4.3.4
discord-py-slash-command==4.2.1
discord-ui==5.1.6
discord.py==2.1.0
discordSuperUtils==0.1.6
dnspython==2.2.1
docopt==0.6.2
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
emoji==2.2.0
feedparser==6.0.10
frozenlist==1.3.3
gevent==22.10.2
greenlet==2.0.1
html5lib==1.1
humanfriendly==10.0
idna==3.4
import-expression==1.1.4
jishaku==2.5.1
line-profiler==4.0.2
MarkupSafe==2.1.1
motor==3.1.1
multidict==6.0.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
pymongo==4.3.3
PyMySQL==1.0.2
PyNaCl==1.5.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyxivapi==0.5.1
redis==4.4.0
requests==2.28.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
spotipy==2.21.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
Werkzeug==2.2.2
yarl==1.8.1
youtube-dl==2021.12.17
zope.event==4.5.0
zope.interface==5.5.2